move audio utils to core
