+++ packages/client/package.json
@@ -76,6 +76,7 @@
     "typescript": "5.8.2",
     "typescript-eslint": "^8.26.0",
     "vite": "^6.2.1",
+    "vite-plugin-node-polyfills": "^0.23.0",
     "vite-tsconfig-paths": "^5.1.4"
   }
 }
