+++ packages/core/src/index.ts
@@ -12,6 +12,7 @@ export * from './roles';
 export * from './runtime';
 export * from './settings';
 export * from './uuid';
+export * from './audioUtils';
 
 // Export service types
 export * from './services';
