+++ packages/docs/packages/clients/twitter.md
@@ -40,7 +40,7 @@ TWITTER_POLL_INTERVAL=120  # Poll interval in seconds
 TWITTER_TARGET_USERS=      # Comma-separated list of target users
 
 # Post Generation Settings
-ENABLE_TWITTER_POST_GENERATION=true
+TWITTER_ENABLE_POST_GENERATION=true
 POST_INTERVAL_MIN=90       # Minimum interval between posts (minutes)
 POST_INTERVAL_MAX=180      # Maximum interval between posts (minutes)
 POST_IMMEDIATELY=false     # Skip approval workflow
