--- packages/plugin-browser/src/services/audioUtils.ts
@@ -1,31 +1,0 @@
-/**
- * Generates a WAV file header based on the provided audio parameters.
- *
- * @param {number} audioLength - The length of the audio data in bytes.
- * @param {number} sampleRate - The sample rate in Hz.
- * @param {number} [channelCount=1] - The number of audio channels (default is 1).
- * @param {number} [bitsPerSample=16] - The number of bits per sample (default is 16).
- * @returns {Buffer} The generated WAV file header as a Buffer.
- */
-export function getWavHeader(
-  audioLength: number,
-  sampleRate: number,
-  channelCount = 1,
-  bitsPerSample = 16
-): Buffer {
-  const wavHeader = Buffer.alloc(44);
-  wavHeader.write('RIFF', 0);
-  wavHeader.writeUInt32LE(36 + audioLength, 4); // Length of entire file in bytes minus 8
-  wavHeader.write('WAVE', 8);
-  wavHeader.write('fmt ', 12);
-  wavHeader.writeUInt32LE(16, 16); // Length of format data
-  wavHeader.writeUInt16LE(1, 20); // Type of format (1 is PCM)
-  wavHeader.writeUInt16LE(channelCount, 22); // Number of channels
-  wavHeader.writeUInt32LE(sampleRate, 24); // Sample rate
-  wavHeader.writeUInt32LE((sampleRate * bitsPerSample * channelCount) / 8, 28); // Byte rate
-  wavHeader.writeUInt16LE((bitsPerSample * channelCount) / 8, 32); // Block align ((BitsPerSample * Channels) / 8)
-  wavHeader.writeUInt16LE(bitsPerSample, 34); // Bits per sample
-  wavHeader.write('data', 36); // Data chunk header
-  wavHeader.writeUInt32LE(audioLength, 40); // Data chunk size
-  return wavHeader;
-}
